requirements.txt
python-fasthtml==0.0.8
aiosqlite
python-multipart